requirements.txt
pynmea2
Buoy-Base